refactor(generator): move `SourceRoots()`
Move this function to the `config`, I am planning to use it outside of
the `parser` package for the Rust+Prost codec.
