+++ generator/internal/config/source_roots.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package parser
+package config
 
 import "strings"
