+++ generator/internal/sidekick/refreshall.go
@@ -26,7 +26,6 @@ import (
 	"sync"
 
 	"github.com/googleapis/google-cloud-rust/generator/internal/config"
-	"github.com/googleapis/google-cloud-rust/generator/internal/parser"
 )
 
 func init() {
@@ -47,7 +46,7 @@ func overrideSources(rootConfig *config.Config) (*config.Config, error) {
 	override := *rootConfig
 	override.Codec = maps.Clone(rootConfig.Codec)
 	override.Source = maps.Clone(rootConfig.Source)
-	for _, root := range parser.SourceRoots(rootConfig.Source) {
+	for _, root := range config.SourceRoots(rootConfig.Source) {
 		configPrefix := strings.TrimSuffix(root, "-root")
 		if _, ok := rootConfig.Source[root]; !ok {
 			continue
